Centralized level map related code into the TDLevelMap class. Documentation included.
